requirements.txt
cs50
Flask
Flask-Session
requests
functools
urllib.parse
tempfile
werkzeug.security
datetime